文件：增加Instrument的Bean，Dao，Biz

+++ src/main/java/team/wy/orchestra/biz/ChurchConcertBiz.java
@@ -83,7 +83,7 @@ public class ChurchConcertBiz {
         return pageCount;
     }
 
-    public ChurchConcert getById(int id) {
+    public ChurchConcert getById(long id) {
         ChurchConcert churchConcert = null;
 
         try {
